// src/shared/ui/Flex/lib/types.ts
export type FlexAlign = "start" | "end" | "center" | "baseline" | "stretch";

export type FlexJustify =
  | "start"
  | "end"
  | "center"
  | "space-between"
  | "space-around"
  | "space-evenly"
  | "stretch";
export type FlexDirection = "row" | "column" | "row-reverse" | "column-reverse";

export type FlexGap =
  | "0-5"
  | "0"
  | "1"
  | "1-5"
  | "2"
  | "2-5"
  | "3"
  | "3-5"
  | "4"
  | "4-5"
  | "5"
  | "6"
  | "7"
  | "8"
  | "9"
  | "10";